Patch handlebars, eliminate `hbs` global variable
