+++ app/utilityFunctionsMostlyText.js
@@ -4,7 +4,7 @@ const fs = require('fs')
 const sharp = require('sharp')
 const path = require('path')
 const sanitizeHtml = require('sanitize-html')
-// const hbs
+const hbs = require('./viewEngine.js')
 const User = require('./models/user')
 
 // these requires are not in server.js bc they're only used here
